Move the server-directory helpers out of constants.py and the RPC integration test into the collected test tree
constants.py grew from a handful of paths to two hundred lines of directory
verification and private-file opens; those now live in _util/server_dir.py
and constants.py only holds the module-level paths built from them.

test_rpc_integration.py sat beside the CLI, where neither the root suite nor
the sandbox-tools CI job collects it, and it launched the CLI through a module
path that only resolves from the package directory. It now lives in the
package's tests/ tree on the shared conftest fixtures, so its socket and
directory permission assertions run in CI.
